refactor: split graphql types
