+++ src/routes/graphql/types/member.ts
@@ -1,0 +1,28 @@
+import { GraphQLEnumType, GraphQLFloat, GraphQLInt, GraphQLNonNull, GraphQLObjectType } from "graphql";
+
+export const memberTypeId = new GraphQLEnumType({
+  name: 'MemberTypeId',
+  values: {
+    BASIC: { value: 'BASIC' },
+    BUSINESS: { value: 'BUSINESS' },
+  },
+});
+
+const memberType = new GraphQLObjectType({
+  name: 'Member',
+  description: 'Member information',
+  fields: () => ({
+    id: {
+      type: memberTypeId,
+    },
+    discount: {
+      type: new GraphQLNonNull(GraphQLFloat),
+    },
+    postsLimitPerMonth: {
+      type: new GraphQLNonNull(GraphQLInt),
+    },
+  }),
+});
+
+
+export {memberType as default}
