+++ src/routes/graphql/types/post.ts
@@ -1,0 +1,20 @@
+import { GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
+import { UUIDType } from "./uuid.js";
+
+const postType = new GraphQLObjectType({
+  name: 'Post',
+  description: "User's post",
+  fields: () => ({
+    id: {
+      type: new GraphQLNonNull(UUIDType),
+    },
+    title: {
+      type: new GraphQLNonNull(GraphQLString),
+    },
+    content: {
+      type: new GraphQLNonNull(GraphQLString),
+    },
+  }),
+});
+
+export {postType as default}
