+++ src/routes/graphql/types/profile.ts
@@ -1,0 +1,33 @@
+import { GraphQLBoolean, GraphQLInt, GraphQLNonNull, GraphQLObjectType } from "graphql";
+import { UUIDType } from "./uuid.js";
+import memberType from "./member.js";
+import { PrismaClient } from "@prisma/client";
+
+const profileType = new GraphQLObjectType({
+  name: 'Profile',
+  description: 'Profile',
+  fields: () => ({
+    id: {
+      type: new GraphQLNonNull(UUIDType),
+    },
+    isMale: {
+      type: new GraphQLNonNull(GraphQLBoolean),
+    },
+    yearOfBirth: {
+      type: new GraphQLNonNull(GraphQLInt),
+    },
+    memberType: {
+      type: new GraphQLNonNull(memberType),
+      resolve: async (profile, _args, context: PrismaClient) => {
+        const memberType = await context.memberType.findUnique({
+          where: {
+            id: profile.memberTypeId,
+          },
+        });
+        return memberType;
+      },
+    },
+  }),
+});
+
+export {profileType as default}
