+++ src/routes/graphql/types/user.ts
@@ -1,0 +1,75 @@
+import { GraphQLFloat, GraphQLList, GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
+import { UUIDType } from "./uuid.js";
+import { PrismaClient } from "@prisma/client";
+import postType from "./post.js";
+import profileType from "./profile.js";
+
+const userType = new GraphQLObjectType({
+  name: 'User',
+  description: 'User',
+  fields: () => ({
+    id: {
+      type: new GraphQLNonNull(UUIDType),
+    },
+    name: {
+      type: new GraphQLNonNull(GraphQLString),
+    },
+    balance: {
+      type: new GraphQLNonNull(GraphQLFloat),
+    },
+    profile: {
+      type: profileType,
+      resolve: async (user, _args, context: PrismaClient) => {
+        const profile = await context.profile.findUnique({
+          where: {
+            userId: user.id,
+          },
+        });
+        return profile;
+      },
+    },
+    posts: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(postType))),
+      resolve: async (user, _args, context: PrismaClient) => {
+        const posts = context.post.findMany({
+          where: {
+            authorId: user.id,
+          },
+        });
+        return posts;
+      },
+    },
+    userSubscribedTo: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(userType))),
+      resolve: async (user, _args, context: PrismaClient) => {
+        const users = context.user.findMany({
+          where: {
+            subscribedToUser: {
+              some: {
+                subscriberId: user.id,
+              },
+            },
+          },
+        });
+        return users;
+      },
+    },
+    subscribedToUser: {
+      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(userType))),
+      resolve: async (user, _args, context: PrismaClient) => {
+        const users = context.user.findMany({
+          where: {
+            userSubscribedTo: {
+              some: {
+                authorId: user.id,
+              },
+            },
+          },
+        });
+        return users;
+      },
+    },
+  }),
+});
+
+export {userType as default}
